+++ tas/include/fastpath.h
@@ -55,7 +55,7 @@ struct qman_thread {
   /************************************/
   /* modified by owner thread */
   struct queue** timewheel;
-  uint64_t timewheel_granularity_ns;
+  uint32_t timewheel_granularity_ns;
   uint32_t timewheel_head_idx;
   uint32_t timewheel_len;
   uint32_t timewheel_count;   /*> Number of flows currently in the timewheel */
